Move block editor specific scripts to block editor feature class
